refactor: rename buscar-ofertas to explorar + change file structure
