+++ frontend/app/(protected)/pago/checkout.tsx
@@ -1,14 +1,14 @@
 import {Button, TextInput, View, Text, Image, TouchableOpacity, StyleSheet, ScrollView, ActivityIndicator} from "react-native";
 import {useContext, useEffect, useState} from "react";
-import CartItem, {ItemData} from "../_components/CartItem";
-import TotalFooter from "../_components/TotalFooter";
+import CartItem, {ItemData} from "../../_components/CartItem";
+import TotalFooter from "../../_components/TotalFooter";
 import {Elements, PaymentElement, useElements, useStripe} from '@stripe/react-stripe-js';
 import {loadStripe, Stripe} from '@stripe/stripe-js';
 import { router, useLocalSearchParams } from "expo-router";
 import { usePayment } from "@/contexts/PaymentContext";
-import { useAuth } from "../../contexts/AuthContext";
-import SuccessModal from "../_components/SuccessModal";
-import { Products } from "../../utils/productDetails";
+import { useAuth } from "../../../contexts/AuthContext";
+import SuccessModal from "../../_components/SuccessModal";
+import { Products } from "../../../utils/productDetails";
 import colors from "@/assets/styles/colors";
 import globalStyles from "@/assets/styles/globalStyles";
 import Ionicons from "@expo/vector-icons/Ionicons";
