+++ frontend/app/_screens/MisOfertasCamionero.tsx
@@ -51,8 +51,8 @@ const MisOfertasCamionero = () => {
                         <Text style={styles.emptySubtitle}>
                             Aún no has solicitado ninguna oferta. Vuelve a la página principal y empieza a explorar oportunidades.
                         </Text>
-                        <TouchableOpacity style={styles.emptyButton} onPress={() => router.push("/buscar-ofertas")}>
-                            <Text style={styles.emptyButtonText}>Buscar Ofertas</Text>
+                        <TouchableOpacity style={styles.emptyButton} onPress={() => router.push("/explorar")}>
+                            <Text style={styles.emptyButtonText}>Explorar Ofertas</Text>
                         </TouchableOpacity>
                     </View>
 
@@ -77,8 +77,8 @@ const MisOfertasCamionero = () => {
                         <Text style={styles.emptySubtitle}>
                             Vuelve a la página principal para buscar ofertas.
                         </Text>
-                        <TouchableOpacity style={styles.emptyButton} onPress={() => router.push("/buscar-ofertas")}>
-                            <Text style={styles.emptyButtonText}>Buscar Ofertas</Text>
+                        <TouchableOpacity style={styles.emptyButton} onPress={() => router.push("/explorar")}>
+                            <Text style={styles.emptyButtonText}>Explorar Ofertas</Text>
                         </TouchableOpacity>
                     </View>
                 );
